+++ src/tabs/_example/index.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import TDemoHeader from '../../../site/mobile/components/DemoHeader';
 import Base from './base';
+import Horizontal from './horizontal';
 import Scroll from './scroll';
 import Vertical from './vetical';
 import Size from './size';
@@ -12,7 +13,8 @@ export default function () {
   return (
     <div className="tdesign-mobile-demo">
       <TDemoHeader title="Tabs 选项卡" summary="用于切换不同场景" />
-      <Base></Base>
+      <Base />
+      <Horizontal></Horizontal>
       <Vertical></Vertical>
       <Scroll></Scroll>
       <Size></Size>
