+++ src/toast/_example/index.jsx
@@ -1,6 +1,9 @@
 import React from 'react';
+import Text from './text';
+import Icon from './icon';
 import Base from './base';
-import Other from './other';
+import IconText from './iconText';
+import Position from './position';
 import Mask from './mask';
 import TDemoHeader from '../../../site/mobile/components/DemoHeader';
 import './style.less';
@@ -9,8 +12,11 @@ export default function () {
   return (
     <div className="tdesign-mobile-demo">
       <TDemoHeader title="Toast 轻提示" />
+      <Text />
+      <Icon />
+      <IconText />
       <Base />
-      <Other />
+      <Position />
       <Mask />
     </div>
   );
